Move everything from under SWF.
