+++ flow.cabal
@@ -20,10 +20,10 @@ source-repository head
   location:          git@github.com:swift-nav/flow.git
 
 library
-  exposed-modules:     Network.AWS.SWF.Flow
-                     , Network.AWS.SWF.Flow.Helper
-  other-modules:       Network.AWS.SWF.Flow.Internal
-                     , Network.AWS.SWF.Flow.Types
+  exposed-modules:     Network.AWS.Flow
+                     , Network.AWS.Flow.Helper
+  other-modules:       Network.AWS.Flow.Internal
+                     , Network.AWS.Flow.Types
   default-language:    Haskell2010
   hs-source-dirs:      src
   ghc-options:         -Wall -fno-warn-orphans
